Extract ProviderOptions to provider_options.h to avoid config.h in smartptrs.h
DeviceInterface::ShapeInitSessionProviderOptions needs ProviderOptions in its
signature. Including config.h from smartptrs.h required making config.h
self-sufficient (12 new standard headers), which caused the C++17 span
incompatibility and downstream lint issues.

Fix: move DeviceFilteringOptions, NamedString, and ProviderOptions to a new
self-sufficient provider_options.h. smartptrs.h includes the lightweight new
header; config.h adds one #include and replaces the three definitions with
transparent type aliases, so all Config::ProviderOptions call sites compile
unchanged.
